Added Order.java file
This class file can contain an array of menu items, and an order id.

We could load up an array list with menu items, and then we can put the array list into this order object, slap an id on it, and then close the package.

Then we could send this "Order" item off to the kitchen.